+++ src/app/core/components/sidenav/sidenav.component.ts
@@ -1,22 +1,18 @@
-import {
-  AfterViewInit,
-  Component,
-  Input,
-  OnDestroy,
-  ViewChild,
-} from '@angular/core';
+import { AfterViewInit, Component, Input, ViewChild } from '@angular/core';
 import { MatDrawer } from '@angular/material/sidenav';
-import { Observable, Subject } from 'rxjs';
+import { Observable } from 'rxjs';
 import { takeUntil, tap } from 'rxjs/operators';
+import { DisposableComponent } from 'src/app/shared/components';
 
 @Component({
   selector: 'sb-sidenav',
   templateUrl: './sidenav.component.html',
   styleUrls: ['./sidenav.component.scss'],
 })
-export class SidenavComponent implements AfterViewInit, OnDestroy {
-  private onDestroy = new Subject<boolean>();
-
+export class SidenavComponent
+  extends DisposableComponent
+  implements AfterViewInit
+{
   @Input()
   public toggleSidenav!: Observable<boolean | undefined>;
 
@@ -30,10 +26,5 @@ export class SidenavComponent implements AfterViewInit, OnDestroy {
         takeUntil(this.onDestroy)
       )
       .subscribe();
-  }
-
-  public ngOnDestroy(): void {
-    this.onDestroy.next(true);
-    this.onDestroy.complete();
   }
 }