+++ src/app/shared/interfaces/basic-entity.interface.ts
@@ -1,0 +1,6 @@
+/** Basic entity that is a base class for all core entities in the app. */
+export interface BasicEntity {
+  id: number;
+  name: string;
+  description?: string;
+}