+++ src/app/watchlist/components/index.ts
@@ -1,0 +1,2 @@
+export * from './watchlist-item-editor-dialog/watchlist-item-editor-dialog.component';
+export * from './watchlist-items-table/watchlist-items-table.component';