+++ src/app/watchlist/interfaces/index.ts
@@ -1,0 +1,5 @@
+export * from './position-size.interface';
+export * from './target-price.interface';
+export * from './watchlist-item.interface';
+export * from './watchlist-list-item.interface';
+export * from './watchlist.interface';